MAIN FILE CLEANUP: Finished. Moved all operations to BooleanOperations.jl, save() moved to smt_representation, and main BooleanSatisfiability.jl file is now just include() and export(). We can put any global definitions here as well.
